+++ projects/js-packages/publicize-components/src/components/form/index.tsx
@@ -6,6 +6,7 @@
  * sharing message.
  */
 
+import { siteHasFeature } from '@automattic/jetpack-script-data';
 import { Disabled, PanelRow } from '@wordpress/components';
 import { Fragment } from '@wordpress/element';
 import useAttachedMedia from '../../hooks/use-attached-media';
@@ -14,11 +15,13 @@ import useMediaDetails from '../../hooks/use-media-details';
 import useMediaRestrictions from '../../hooks/use-media-restrictions';
 import usePublicizeConfig from '../../hooks/use-publicize-config';
 import useSocialMediaConnections from '../../hooks/use-social-media-connections';
+import { features } from '../../utils';
 import { ThemedConnectionsModal as ManageConnectionsModal } from '../manage-connections-modal';
 import { SocialPostModal } from '../social-post-modal/modal';
 import { ConnectionsList } from './connections-list';
 import { EmptyState } from './empty-state';
 import { EnhancedFeaturesNudge } from './enhanced-features-nudge';
+import { PreviewPostsTrigger } from './preview-posts-trigger';
 import { SharePostForm } from './share-post-form';
 
 /**
@@ -59,7 +62,11 @@ export default function PublicizeForm() {
 			<EmptyState />
 			{ hasConnections ? (
 				<>
-					<SocialPostModal />
+					{ siteHasFeature( features.UNIFIED_UI_V1 ) ? (
+						<PreviewPostsTrigger />
+					) : (
+						<SocialPostModal />
+					) }
 					<EnhancedFeaturesNudge />
 				</>
 			) : null }