+++ projects/js-packages/publicize-components/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.1] - 2025-12-01
+### Added
+- Add unified UI v1 feature flag. [#46061]
+- Create a hook for getting the Social Image Generator preview [#46100]
+
+### Changed
+- Updated Jetpack sidebar labels for post editor sidebar. [#46098]
+- Update package dependencies. [#46143]
+
 ## [1.7.0] - 2025-11-24
 ### Changed
 - Add title for the Facebook share preview. [#45938]
@@ -1431,6 +1440,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated package dependencies. [#24470]
 
+[1.7.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.6.1...v1.7.0
 [1.6.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.6.0...v1.6.1
 [1.6.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.5.1...v1.6.0